django_server/requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.4.0
APScheduler==3.10.4
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
automium==0.2.6
automium_web==0.1.1
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.6
django-admin-interface==0.28.8
django-colorfield==0.11.0
django-flat-responsive==2.0
djangorestframework==3.15.2
drf-yasg==1.21.7
Flask==3.0.3
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
inflection==0.5.1
itsdangerous==2.2.0
itypes==1.2.0
Jinja2==3.1.4
legacy==0.1.7
magic-filter==1.0.12
MarkupSafe==2.1.5
multidict==6.0.5
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
setuptools==71.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
text-unidecode==1.3
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.3
yarl==1.9.4
